Add new ChallengeSerializerTests
This commit is adding a new test class, ChallengeSerializerTests into the common functionalities. The newly introduced class is useful in testing the Challenge.Serializer's deserialize capability to validate if given JSON object is successfully translated back into the Challenge instance. Ensuring this functionality is critical to ensure smooth data management for the application's challenge progress tracking.